# trie: avoid rlp-encoding when trie hash is already available